Make asset-resolver into a subcrate
