+++ aw-server/src/main.rs
@@ -145,7 +145,7 @@ async fn main() -> Result<(), rocket::Error> {
         // Even if legacy_import is set to true it is disabled on Android so
         // it will not happen there
         datastore: Mutex::new(aw_datastore::Datastore::new(db_path, legacy_import)),
-        asset_resolver: Box::new(ProjectAssetResolver),
+        asset_resolver: Box::new(asset_resolver::ProjectAssetResolver),
         device_id,
     };
